+++ apps/web/src/boot.ts
@@ -1,4 +1,6 @@
 import "./style.css";
+import { injectSiteFooterMeta } from "./site-meta";
 import { initNav } from "./site-nav";
 
 initNav();
+injectSiteFooterMeta();